Extract workspace engine and upload helpers
